🐛 Fix flashing wallet bug and errors related to myAlgo signing
Cannot set localStorage on every query update because when signing myAlgo transactions the popUp window can throw errors.

Commented out the entire useEffect because it can cause a flashing wallet to appear and disappear. With the addition of the disconnect buttons we are constraining disconnect events to Algodex, not the wallet clients, therefore, this useEffect check is not needed anymore.